requirements.txt
requests==2.31.0
aiogram==2.25.1
python-dotenv==1.0.0